Refactor wz-table directive
