+++ public/directives/wz-table/index.js
@@ -1,0 +1,13 @@
+/*
+ * Wazuh app - Wazuh table directive
+ * Copyright (C) 2018 Wazuh, Inc.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Find more information about this on the LICENSE file.
+ */
+import './wz-table-filter-service';
+import './wz-table-directive';
