+++ public/directives/wz-table/wz-table.html
@@ -18,11 +18,11 @@
             <tr ng-class="allowClick ? 'cursor-pointer' : ''" class="wz-word-wrap" ng-repeat="item in pagedItems[currentPage] | filter:{item:'!'}" ng-click="clickAction(item)">
                 <td ng-repeat="key in keys" ng-if="path !== '/decoders' && path !== '/rules'">
                     {{
-                        nonDecoderValue(key,item) | limitTo: extraLimit ? 100 : 25
+                        parseValue(key,item) | limitTo: extraLimit ? 100 : 25
                     }}
-                    <span ng-if="nonDecoderValue(key,item).length > (extraLimit ? 100 : 25)">...</span>
-                    <md-tooltip ng-if="nonDecoderValue(key,item).length > (extraLimit ? 100 : 25)" md-direction="bottom" class="wz-tooltip">
-                        {{ nonDecoderValue(key,item) }}
+                    <span ng-if="parseValue(key,item).length > (extraLimit ? 100 : 25)">...</span>
+                    <md-tooltip ng-if="parseValue(key,item).length > (extraLimit ? 100 : 25)" md-direction="bottom" class="wz-tooltip">
+                        {{ parseValue(key,item) }}
                     </md-tooltip>
                 </td>
                 <td ng-repeat="key in keys" ng-if="path === '/rules'" ng-click="key === 'id' && searchRuleId($event,item.id)">
@@ -35,21 +35,21 @@
 
                     <span ng-if="key !== 'id'">
                         {{
-                            nonDecoderValue(key,item) | limitTo: extraLimit ? 100 : 25
+                            parseValue(key,item) | limitTo: extraLimit ? 100 : 25
                         }}
                     </span>
-                    <span ng-if="nonDecoderValue(key,item).length > (extraLimit ? 100 : 25)">...</span>
-                    <md-tooltip ng-if="nonDecoderValue(key,item).length > (extraLimit ? 100 : 25)" md-direction="bottom" class="wz-tooltip">
-                        {{ nonDecoderValue(key,item) }}
+                    <span ng-if="parseValue(key,item).length > (extraLimit ? 100 : 25)">...</span>
+                    <md-tooltip ng-if="parseValue(key,item).length > (extraLimit ? 100 : 25)" md-direction="bottom" class="wz-tooltip">
+                        {{ parseValue(key,item) }}
                     </md-tooltip>
                 </td>
                 <td ng-repeat="key in keys" ng-if="path === '/decoders'">
                     {{
-                        decoderValue(key,item) | limitTo: extraLimit ? 100 : 25
+                        parseValue(key,item) | limitTo: extraLimit ? 100 : 25
                     }}
-                    <span ng-if="decoderValue(key,item).length > (extraLimit ? 100 : 25)">...</span>
-                    <md-tooltip ng-if="decoderValue(key,item).length > (extraLimit ? 100 : 25)" md-direction="bottom" class="wz-tooltip">
-                        {{ decoderValue(key,item) }}
+                    <span ng-if="parseValue(key,item).length > (extraLimit ? 100 : 25)">...</span>
+                    <md-tooltip ng-if="parseValue(key,item).length > (extraLimit ? 100 : 25)" md-direction="bottom" class="wz-tooltip">
+                        {{ parseValue(key,item) }}
                     </md-tooltip>
                 </td>
                 <td ng-if="path === '/agents'" ng-click="$event.stopPropagation()" class="cursor-default">
